Update Semester Project
We were asked to implement ASCII codes in the project somehow, and the first 22 lines are all used in the usage of ASCII code for the interface. Other than that switches, loops, if else and more functions were used, string, char, bool, float were also used to showcase in depth understanding of all data types pointers were not used to avoid making it too complicated, I wanted the project to be fully functional, easy to use and user friendly as well as having a neat interface.